Bug 1707571 Fixup to fxa_auth_events isverified field
Fixup to #1975

+++ sql/moz-fx-data-shared-prod/firefox_accounts_derived/fxa_auth_events_v1/query.sql
@@ -12,7 +12,7 @@ WITH base AS (
               SELECT AS STRUCT
                 jsonPayload.fields.* EXCEPT (device_id, user_id) REPLACE(
                   -- See https://bugzilla.mozilla.org/show_bug.cgi?id=1707571
-                  CAST(NULL AS FLOAT) AS isverified
+                  CAST(NULL AS FLOAT64) AS isverified
                 ),
                 TO_HEX(SHA256(jsonPayload.fields.user_id)) AS user_id
             ) AS fields
